Type: starting Dog classification repository
The Dog classification repository is related to an UDACITY project.
Throughout this project we must look at various type of pictures
and ultimately determined whether it is a human or a dog. If it is
a dog we must tell what breed it is, and if it is a human, to what
do breed it looks like