Subject: Update of the CNN Model architecture
Completly modified the Model architecture as teh previous one only
reached 2% of accuracy on the test. The updates consist of:
- Change Relu to Relu6
- Add BatchNorm2d of the Convolutional layers
- change the CNN with L1 3->32, L2 32->64, L3 64->128
- Removed 1 hidden layer and reduce the size of the output of the h1
to 500
- Note1: I had a issue with the size of the model.pt which is saved if
The validation loss is better than the previous epoch, but the size was
too big and the memory crashed. So I reduce this specific element
self.fc1 = nn.Linear(128 * 14 * 14, 500)
 - Note2: put print(x.shape) before the x = x.view(-1, 128 * 14 * 14)
 to make sure you have the rigth sizing from the CNN to the MLP