Subject: add Torchsummary
Torch Summary is a very good tool to give a you a quick overlook of
your architecture model. It replaces the nice recap that is avaialble
on Keras and not on pytorch. I strongly recommend to use it. Just
need to install it via the terminal `pip install torchsummary`.
On this model architecture, it helps me to better understand the
structure and with additional print(x.shape) at various points in the
fwd pass (which I removed for better reading), it gives a clear picture.