Subject: final submission of dog classification project
This commit included a cleaning of all the import in the various cells
as all the imports and reference were already added at the beginning
of the code in a previous commit. With these last commit, the
program should work smoothly, but I will have to adjust some path that
were not the same one on the workspace I used when I modified the code.
Also, there are several topics that I will need to work on:
1. Manage several faces in the same picture
2. Migrate the solution to flask
3. Re-use the trained model .pt
4. improve model once a GPU available